requirements.txt
Django
djangorestframework
pymongo
gunicorn
dj-static
django-cors-headers
beautifulsoup4
requests
